requirements.txt
Flask==2.3.2
requests==2.31.0
APScheduler==3.11.1
python-dotenv==1.0.0
Pillow==10.0.0
gunicorn==20.1.0
SQLAlchemy==2.0.19
alembic==1.11.1

